# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CUSIT University Event Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  faculty: "faculty",
  coordinator: "coordinator",
  society_head: "society_head",
  admin: "admin",
} as const;

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type EventCategory = (typeof EventCategory)[keyof typeof EventCategory];

export const EventCategory = {
  seminar: "seminar",
  workshop: "workshop",
  competition: "competition",
  sports: "sports",
  society: "society",
  cultural: "cultural",
  technical: "technical",
} as const;

export type AttendanceStatus =
  (typeof AttendanceStatus)[keyof typeof AttendanceStatus];

export const AttendanceStatus = {
  absent: "absent",
  present: "present",
  late: "late",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  department?: string | null;
  registrationNumber?: string | null;
  avatarUrl?: string | null;
}

export interface LoginRequest {
  userId: number;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  category: EventCategory;
  status: EventStatus;
  organizerId: number;
  organizerName: string;
  organizerRole?: UserRole;
  startsAt: string;
  endsAt: string;
  venue: string;
  capacity: number;
  registrationCount: number;
  bannerUrl?: string | null;
  rejectionReason?: string | null;
  createdAt?: string;
}

export type EventDetail = Event & {
  isRegistered?: boolean;
  myAttendanceStatus?: AttendanceStatus;
  hasCertificate?: boolean;
};

export interface EventCreatePayload {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  category: EventCategory;
  startsAt: string;
  endsAt: string;
  venue: string;
  /** @minimum 1 */
  capacity: number;
  bannerUrl?: string | null;
}

export interface EventUpdatePayload {
  title?: string;
  description?: string;
  category?: EventCategory;
  startsAt?: string;
  endsAt?: string;
  venue?: string;
  capacity?: number;
  bannerUrl?: string | null;
  status?: EventStatus;
}

export type EventApprovalPayloadDecision =
  (typeof EventApprovalPayloadDecision)[keyof typeof EventApprovalPayloadDecision];

export const EventApprovalPayloadDecision = {
  approve: "approve",
  reject: "reject",
} as const;

export interface EventApprovalPayload {
  decision: EventApprovalPayloadDecision;
  rejectionReason?: string | null;
}

export interface Registration {
  id: number;
  eventId: number;
  userId: number;
  userName: string;
  userEmail?: string;
  userDepartment?: string | null;
  registeredAt: string;
}

export interface RegistrationWithEvent {
  id: number;
  event: Event;
  registeredAt: string;
  attendanceStatus?: AttendanceStatus;
  hasCertificate?: boolean;
}

export interface AttendanceRow {
  registrationId: number;
  userId: number;
  userName: string;
  userEmail?: string;
  userDepartment?: string | null;
  status: AttendanceStatus;
  markedAt?: string | null;
}

export type AttendanceMarkPayloadEntriesItem = {
  registrationId: number;
  status: AttendanceStatus;
};

export interface AttendanceMarkPayload {
  entries: AttendanceMarkPayloadEntriesItem[];
}

export interface IssueCertificatesResult {
  issuedCount: number;
}

export interface Certificate {
  id: number;
  eventId: number;
  userId: number;
  code: string;
  issuedAt: string;
}

export interface CertificateWithEvent {
  id: number;
  code: string;
  issuedAt: string;
  event: Event;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  body: string;
  eventId?: number | null;
  createdAt: string;
  isRead: boolean;
}

export interface Feedback {
  id: number;
  eventId: number;
  userId: number;
  userName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  createdAt: string;
}

export interface FeedbackPayload {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
}

export interface DashboardStats {
  totalEvents: number;
  upcomingEvents: number;
  pendingApprovals: number;
  totalRegistrations: number;
  totalCertificates: number;
  activeUsers: number;
  avgFeedbackRating?: number;
}

export interface CategoryCount {
  category: EventCategory;
  count: number;
}

export interface TrendPoint {
  date: string;
  count: number;
}

export interface TopEvent {
  eventId: number;
  title: string;
  category?: EventCategory;
  registrationCount: number;
  attendanceRate: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  event_created: "event_created",
  event_approved: "event_approved",
  event_rejected: "event_rejected",
  registration: "registration",
  certificate: "certificate",
  feedback: "feedback",
} as const;

export interface ActivityItem {
  type: ActivityItemType;
  message: string;
  eventId?: number | null;
  timestamp: string;
}

export interface EventReport {
  event: Event;
  registrationCount: number;
  attendedCount: number;
  absentCount: number;
  certificatesIssued: number;
  avgRating: number;
  feedbackCount: number;
}

export type ListUsersParams = {
  role?: UserRole;
};

export type ListEventsParams = {
  status?: EventStatus;
  category?: string;
  search?: string;
  organizerId?: number;
  upcoming?: boolean;
};

export type GetRegistrationsTrendParams = {
  days?: number;
};

export type GetTopEventsParams = {
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};