# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CUSIT University Event Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginAs
      tags: [auth]
      summary: Login as a seeded user (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events with filters
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/EventStatus" } }
        - { name: category, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: organizerId, in: query, schema: { type: integer } }
        - { name: upcoming, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
    post:
      operationId: createEvent
      tags: [events]
      summary: Propose a new event (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventCreatePayload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventDetail" }
    patch:
      operationId: updateEvent
      tags: [events]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventUpdatePayload" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
    delete:
      operationId: deleteEvent
      tags: [events]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /events/{id}/approve:
    post:
      operationId: approveEvent
      tags: [events]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventApprovalPayload" }
      responses:
        "200":
          description: Approved or rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
...
  /events/{id}/attendance:
    get:
      operationId: listEventAttendance
      tags: [attendance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Attendance roster
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AttendanceRow" }
    post:
      operationId: markAttendance
      tags: [attendance]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendanceMarkPayload" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AttendanceRow" }
  /events/{id}/certificates/issue:
    post:
      operationId: issueCertificates
      tags: [certificates]
      summary: Issue certificates to all attended students
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Certificates issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssueCertificatesResult" }
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [student, faculty, coordinator, society_head, admin]
    EventStatus:
      type: string
      enum: [pending, approved, rejected, completed, cancelled]
    EventCategory:
      type: string
      enum: [seminar, workshop, competition, sports, society, cultural, technical]
    AttendanceStatus:
      type: string
      enum: [absent, present, late]
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        department: { type: string, nullable: true }
        registrationNumber: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
    LoginRequest:
      type: object
      required: [userId]
      properties:
        userId: { type: integer }
    Event:
      type: object
      required: [id, title, description, category, status, organizerId, organizerName, startsAt, endsAt, venue, capacity, registrationCount]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { $ref: "#/components/schemas/EventCategory" }
        status: { $ref: "#/components/schemas/EventStatus" }
        organizerId: { type: integer }
        organizerName: { type: string }
        organizerRole: { $ref: "#/components/schemas/UserRole" }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        venue: { type: string }
        capacity: { type: integer }
        registrationCount: { type: integer }
        bannerUrl: { type: string, nullable: true }
        rejectionReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            isRegistered: { type: boolean }
            myAttendanceStatus: { $ref: "#/components/schemas/AttendanceStatus" }
            hasCertificate: { type: boolean }
    EventCreatePayload:
      type: object
      required: [title, description, category, startsAt, endsAt, venue, capacity]
      properties:
        title: { type: string, minLength: 3 }
        description: { type: string, minLength: 10 }
        category: { $ref: "#/components/schemas/EventCategory" }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        venue: { type: string }
        capacity: { type: integer, minimum: 1 }
        bannerUrl: { type: string, nullable: true }
    EventUpdatePayload:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { $ref: "#/components/schemas/EventCategory" }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        venue: { type: string }
        capacity: { type: integer }
        bannerUrl: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/EventStatus" }
    EventApprovalPayload:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approve, reject] }
        rejectionReason: { type: string, nullable: true }
...
    AttendanceRow:
      type: object
      required: [registrationId, userId, userName, status]
      properties:
        registrationId: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        userEmail: { type: string }
        userDepartment: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/AttendanceStatus" }
        markedAt: { type: string, format: date-time, nullable: true }
    AttendanceMarkPayload:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            type: object
            required: [registrationId, status]
            properties:
              registrationId: { type: integer }
              status: { $ref: "#/components/schemas/AttendanceStatus" }
    IssueCertificatesResult:
      type: object
      required: [issuedCount]
      properties:
        issuedCount: { type: integer }
...